refactor transformation engine
